#209 children connection for post objects.
This adds a child{PluralName} to hierarchical post_types.

For example:

{
  pages {
    edges {
      node {
        id
        title
        childPages {
          edges {
            node {
              id
              title
              date
            }
          }
        }
      }
    }
  }
}

This isn’t a generic “children” field, as children _can_ be of any type (posts can have parents of different post_types), so we’ll have to think that through a bit more as it will need to be a union of some sort. . .