#296 - Conflict with WooCommerce (and other Custom Post Types with "show_in_rest" set to true).
In the latest release we defaulted to have "show_in_graphql" be true for Post Types and Taxonomies that also had "show_in_rest" set to true, but we didn't properly test this feature as it was causing things to break. So this removes that and only includes post types and taxonomies in the Schema that have `show_in_graphql` set to true.

Also fixes some missing capitalizations with Enums and other misc types related to the Capitalized types issue.