Create WPSchema config variable
PHPStorm complains that the $schema->config[] variable is private at https://github.com/wp-graphql/wp-graphql/blob/develop/src/Utils/InstrumentSchema.php#L53

The warning is caused by the IDE assuming a connection to the webonyx/Type/Schema->config[] rather than the explicitly defined \WPGraphQL\WPSchema object.

The misconnection was due to the config variable not being created in WPSchema.

I added the definition and the constructor's instantiation of the variable in \WPGraphQL\WPSchema.php.

NOTE: It didn't seem to cause any runtime errors; it was just an IDE warning that I wanted to remove. Hopefully it will avoid confusion in the future (and, who knows, it may even avoid runtime errors).